refactor : Task ダイアログ state machine を useTaskDialogState に共通化する
TaskListWithDialogs と KanbanBoard で重複していた行クリック state +
useTaskDetail + EditTaskDialog + DeleteTaskDialog の連動ロジックを
hook に集約する。

- hooks/useTaskDialogState.ts を新設
  - selectedTaskId / deleteOpen の state
  - useTaskDetail と連動した task / isLoading / isError
  - select / openDelete / setDeleteOpen / closeAll の操作
- TaskListWithDialogs.tsx と KanbanBoard.tsx を本 hook 経由に書き換え
- 既存テスト 368 件はそのまま通る

Refs: #256
